backend/requirements.txt
# ========================================
# AI教育助手后端项目依赖
# ========================================

# Web框架和服务器
fastapi==0.104.1              # 现代化Python Web框架
uvicorn[standard]==0.24.0     # ASGI服务器，包含标准依赖

# 数据验证和配置
pydantic==2.5.0               # 数据验证和序列化
pydantic-settings==2.1.0     # 配置管理

# HTTP和网络
python-multipart==0.0.6      # 文件上传支持
requests==2.31.0             # HTTP客户端库（用于外部API调用）

# 日期时间处理
python-dateutil==2.8.2       # 日期时间工具

# 环境变量和配置
python-dotenv==1.0.0         # .env文件支持

# 数据验证
pydantic==2.5.0
pydantic-settings==2.1.0
mypy==1.7.1                   # 类型检查

# 测试框架
pytest==7.4.3                # 测试框架
pytest-asyncio==0.21.1       # 异步测试支持
httpx==0.25.2                # 异步HTTP客户端（用于测试）

# 测试
pytest==7.4.3                # 测试框架
pytest-asyncio==0.21.1       # 异步测试支持
httpx==0.25.2                # 异步HTTP客户端（用于测试）

# 数据库
sqlalchemy==2.0.23
aiosqlite==0.19.0
asyncpg==0.29.0

# 机器学习相关
scikit-learn==1.3.0          # 机器学习库，包含sklearn.metrics.pairwise
numpy                 # 数值计算基础库
scipy                   # 科学计算库

#功能模块
jieba==0.42.1                 # 中文分词
pdfminer.six 

sentence_transformers
# ========================================
# 安装说明：
# pip install -r requirements.txt
#
# 开发环境安装（包含所有依赖）：
# pip install -r requirements.txt
#
# 生产环境安装（排除开发工具）：
# pip install fastapi uvicorn[standard] pydantic pydantic-settings python-multipart requests python-dateutil python-dotenv
# ========================================





